refactor: share common types between client and server
