+++ src/server/routes/lights.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 import Tradfri from '../services/Tradfri';
 import { operateLightOrRoom } from '../services/tradfri-controller';
-import type { UpdateLightOperation } from '../../server/types';
+import type { UpdateLightOperation } from 'src/common';
 import type { LightOperation } from 'node-tradfri-client';
 
 const router = express.Router();
@@ -20,7 +20,7 @@ router.put('/:light', async (req, res) => {
   return res.json(result);
 });
 
-router.get('/:light', async (req, res) => {
+router.get('/:light', async (req) => {
   const tradfri = await Tradfri.getInstance();
   const { light } = req.params;
   return tradfri.getLightInfo(+light);
